refactor(go): split 2 large files + extract parseIssueURL
agent/manager.go (700→200 lines) split into:
- manager_run.go: Run, buildCommand, gateProvider, normalizeProvider/Model
- manager_control.go: Find/Kill/Stop/Get agents, Shutdown

app_reviews.go (730→297 lines) split into:
- app_reviews_inbound.go: inbound review task mgmt
- app_reviews_fix.go: handlePRIssue, handleAutoMerge, conflictPrompt

workflow: replace private parseIssueURL with github.ParseIssueURL;
remove duplicate TestParseIssueURL (covered in github/client_test.go)
