+++ internal/agent/manager.go
@@ -4,16 +4,11 @@ import (
 	"context"
 	"fmt"
 	"log/slog"
-	"os"
-	"regexp"
-	"strings"
 	"sync"
 	"time"
 
-	"github.com/Automaat/sybra/internal/events"
 	"github.com/Automaat/sybra/internal/metrics"
 	"github.com/Automaat/sybra/internal/provider"
-	"github.com/google/uuid"
 )
 
 type EmitFunc func(event string, data any)
@@ -203,498 +198,3 @@ func (m *Manager) RunningCount() int {
 	defer m.mu.RUnlock()
 	return m.liveCount
 }
-
-func (m *Manager) StartAgent(taskID, taskTitle, mode, prompt, dir string, allowedTools []string) (*Agent, error) {
-	return m.Run(RunConfig{TaskID: taskID, Name: taskTitle, Mode: mode, Prompt: prompt, AllowedTools: allowedTools, Dir: dir})
-}
-
-func (m *Manager) Run(cfg RunConfig) (*Agent, error) {
-	// Hard guard: every agent must run in an explicit, existing directory.
-	// An empty Dir means the spawned process inherits Sybra's cwd, which in
-	// dev mode is the Sybra source repo — agents would then mutate its
-	// branches via git checkout. Reject rather than silently leak.
-	if strings.TrimSpace(cfg.Dir) == "" {
-		return nil, fmt.Errorf("agent.Run: Dir is required (empty Dir would leak agent process into Sybra cwd)")
-	}
-	if info, err := os.Stat(cfg.Dir); err != nil {
-		return nil, fmt.Errorf("agent.Run: Dir %q not accessible: %w", cfg.Dir, err)
-	} else if !info.IsDir() {
-		return nil, fmt.Errorf("agent.Run: Dir %q is not a directory", cfg.Dir)
-	}
-
-	resolvedProvider, gateErr := m.gateProvider(cfg)
-	if gateErr != nil {
-		return nil, gateErr
-	}
-
-	if cfg.BashTimeoutMs == 0 {
-		m.mu.RLock()
-		cfg.BashTimeoutMs = m.bashTimeoutMs
-		m.mu.RUnlock()
-	}
-
-	id := uuid.NewString()[:8]
-	ctx, cancel := context.WithCancel(m.ctx)
-
-	now := time.Now().UTC()
-	a := &Agent{
-		ID:          id,
-		TaskID:      cfg.TaskID,
-		Name:        cfg.Name,
-		Mode:        cfg.Mode,
-		Provider:    resolvedProvider,
-		Model:       normalizeModel(resolvedProvider, cfg.Model),
-		Prompt:      cfg.Prompt,
-		State:       StateRunning,
-		StartedAt:   now,
-		LastEventAt: now,
-		cancel:      cancel,
-		sessionCWD:  cfg.Dir,
-		MaxTurns:    cfg.MaxTurns,
-	}
-	if cfg.ResumeSessionID != "" {
-		a.SetSessionID(cfg.ResumeSessionID)
-	}
-	if cfg.Mode == "headless" || cfg.Mode == "interactive" {
-		a.done = make(chan struct{})
-	}
-	if cfg.Mode == "headless" {
-		a.escalationCh = make(chan bool, 1)
-	}
-
-	m.mu.Lock()
-	if !cfg.IgnoreConcurrencyLimit && m.maxConcurrent > 0 && m.liveCount >= m.maxConcurrent {
-		m.mu.Unlock()
-		cancel()
-		return nil, fmt.Errorf("max concurrent agents reached (%d)", m.maxConcurrent)
-	}
-	m.agents[id] = a
-	if a.done != nil {
-		m.liveCount++
-	}
-	m.mu.Unlock()
-
-	metrics.AgentStarted(a.Provider, a.Mode)
-	m.logger.Info("agent.start", "id", id, "taskID", cfg.TaskID, "mode", cfg.Mode, "provider", a.Provider, "model", a.Model)
-
-	switch cfg.Mode {
-	case "headless":
-		go m.runHeadless(ctx, a, cfg)
-	case "interactive":
-		if a.Provider == "codex" {
-			a.promptCh = make(chan string, 1)
-			go m.runCodexConversational(ctx, a, cfg)
-		} else {
-			a.approvalCh = make(chan ApprovalResponse, 1)
-			go m.runConversational(ctx, a, cfg)
-		}
-	default:
-		cancel()
-		return nil, fmt.Errorf("unknown mode: %s", cfg.Mode)
-	}
-
-	m.emit(events.AgentState(id), a)
-	return a, nil
-}
-
-func (m *Manager) markAgentDone(a *Agent) {
-	if a == nil || a.done == nil {
-		return
-	}
-	a.doneOnce.Do(func() {
-		close(a.done)
-		m.mu.Lock()
-		if m.liveCount > 0 {
-			m.liveCount--
-		}
-		m.mu.Unlock()
-	})
-}
-
-func (m *Manager) buildCommand(cfg RunConfig) (string, error) {
-	prov := m.providerForRun(cfg.Provider)
-	if cfg.Model != "" && !safeArgRe.MatchString(cfg.Model) {
-		return "", fmt.Errorf("invalid model %q: must match %s", cfg.Model, safeArgRe)
-	}
-	for _, tool := range cfg.AllowedTools {
-		if !safeArgRe.MatchString(tool) {
-			return "", fmt.Errorf("invalid tool %q: must match %s", tool, safeArgRe)
-		}
-	}
-	model := normalizeModel(prov, cfg.Model)
-	switch prov {
-	case "codex":
-		return buildCodexCommand(model, cfg.RequirePermissions, cfg.Mode == "headless"), nil
-	default:
-		return buildClaudeCommand(model, cfg.AllowedTools, cfg.RequirePermissions), nil
-	}
-}
-
-// buildClaudeCommand builds the display command string for a Claude agent.
-func buildClaudeCommand(model string, allowedTools []string, requirePerms bool) string {
-	parts := []string{"claude"}
-	if len(allowedTools) > 0 {
-		parts = append(parts, "--allowedTools", strings.Join(allowedTools, ","))
-	} else if !requirePerms {
-		parts = append(parts, "--dangerously-skip-permissions")
-	}
-	if model != "" {
-		parts = append(parts, "--model", model)
-	}
-	return strings.Join(parts, " ")
-}
-
-// buildCodexCommand builds the display command string for a Codex agent.
-func buildCodexCommand(model string, requirePerms, headless bool) string {
-	parts := []string{"codex", "exec", "--json", "--skip-git-repo-check", "--ignore-user-config", "--ignore-rules"}
-	parts = append(parts, codexSandboxArgs(requirePerms, headless)...)
-	if model != "" {
-		parts = append(parts, "--model", model)
-	}
-	return strings.Join(parts, " ")
-}
-
-// codexSandboxArgs returns the sandbox/permission flags for `codex exec`.
-// When SYBRA_DISABLE_CODEX_SANDBOX=1 is set, the bwrap-backed sandbox is
-// replaced with `--sandbox danger-full-access`. Required when running
-// inside a Docker/LXC container whose kernel blocks unprivileged user
-// namespaces (kernel.unprivileged_userns_clone=0), where bwrap crashes
-// before the agent can execute any command.
-//
-// headless must be true for headless runs. --sandbox workspace-write asks
-// for user approval on writes outside the workspace; in headless mode there
-// is no UI to serve those approval prompts, so they auto-reject and the run
-// fails. Bypass mode is used instead.
-func codexSandboxArgs(requirePerms, headless bool) []string {
-	if os.Getenv("SYBRA_DISABLE_CODEX_SANDBOX") == "1" {
-		return []string{"--sandbox", "danger-full-access"}
-	}
-	if !requirePerms || headless {
-		// Bypass all approval prompts and sandbox restrictions.
-		// For headless runs this is required regardless of RequirePermissions:
-		// --sandbox workspace-write auto-rejects approval requests since there
-		// is no TTY/UI to serve them, which silently breaks the agent run.
-		return []string{"--dangerously-bypass-approvals-and-sandbox"}
-	}
-	return []string{"--sandbox", "workspace-write"}
-}
-
-// gateProvider resolves the run's provider through the health gate. If the
-// configured provider is unhealthy and auto-failover can supply a healthy
-// peer, the peer is returned. Otherwise returns a typed UnhealthyError so
-// callers can detect via errors.Is(err, provider.ErrProviderUnhealthy).
-func (m *Manager) gateProvider(cfg RunConfig) (string, error) {
-	resolved := m.providerForRun(cfg.Provider)
-	if cfg.IgnoreHealthGate {
-		return resolved, nil
-	}
-	m.mu.RLock()
-	g := m.gate
-	m.mu.RUnlock()
-	if g == nil {
-		return resolved, nil
-	}
-	if g.IsHealthy(resolved) {
-		return resolved, nil
-	}
-	if alt := g.Failover(resolved); alt != "" {
-		metrics.AgentFailover(resolved, alt)
-		m.logger.Warn("agent.run.failover", "from", resolved, "to", alt, "task", cfg.TaskID, "reason", g.Reason(resolved))
-		return alt, nil
-	}
-	reason := g.Reason(resolved)
-	metrics.AgentGated(resolved, reason)
-	return "", &provider.UnhealthyError{
-		Provider: resolved,
-		Reason:   reason,
-	}
-}
-
-func (m *Manager) providerForRun(name string) string {
-	m.mu.RLock()
-	def := m.defaultProv
-	m.mu.RUnlock()
-	if name == "" {
-		name = def
-	}
-	return normalizeProvider(name)
-}
-
-func normalizeProvider(name string) string {
-	switch strings.ToLower(strings.TrimSpace(name)) {
-	case "", "claude":
-		return "claude"
-	case "codex":
-		return "codex"
-	default:
-		return "claude"
-	}
-}
-
-func normalizeModel(prov, model string) string {
-	switch normalizeProvider(prov) {
-	case "codex":
-		switch strings.TrimSpace(model) {
-		case "", "sonnet", "opus":
-			return "gpt-5.4"
-		case "haiku":
-			return "gpt-5.4-mini"
-		default:
-			return model
-		}
-	default:
-		if strings.TrimSpace(model) == "" {
-			return "sonnet"
-		}
-		return model
-	}
-}
-
-// SendPromptToAgent delivers a follow-up prompt to an interactive agent.
-func (m *Manager) SendPromptToAgent(agentID, text string) error {
-	a, err := m.GetAgent(agentID)
-	if err != nil {
-		return err
-	}
-	if a.GetState() == StateStopped {
-		return fmt.Errorf("agent %s is stopped", agentID)
-	}
-
-	// Conversational agents: write to stdin via SendMessage.
-	a.stdinMu.Lock()
-	hasStdin := a.stdinPipe != nil
-	a.stdinMu.Unlock()
-	if hasStdin {
-		return m.SendMessage(agentID, text)
-	}
-
-	// Codex conversational agents: deliver via promptCh.
-	a.mu.RLock()
-	hasCh := a.promptCh != nil
-	a.mu.RUnlock()
-	if hasCh {
-		return m.sendCodexPrompt(agentID, text)
-	}
-
-	return fmt.Errorf("agent %s has no active transport (no stdin pipe or prompt channel)", agentID)
-}
-
-// isLive reports whether an agent is still alive from the user's perspective.
-// Conversational agents switch to StatePaused while idle between turns; they
-// must still be findable so a follow-up prompt can be delivered without
-// spawning a new session.
-func isLive(s State) bool {
-	return s == StateRunning || s == StatePaused
-}
-
-// FindRunningAgentForTask returns the first live agent for the given task
-// matching the provided role. Returns nil if none found. "Live" includes
-// paused conversational agents that are idle-waiting for a follow-up prompt.
-func (m *Manager) FindRunningAgentForTask(taskID string, role Role) *Agent {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-	for _, a := range m.agents {
-		if a.TaskID != taskID || !isLive(a.GetState()) {
-			continue
-		}
-		if RoleFromName(a.Name) != role {
-			continue
-		}
-		return a
-	}
-	return nil
-}
-
-// FindAllRunningAgentsForTask returns all live agents for the given task
-// matching the provided role. An empty role matches all roles.
-func (m *Manager) FindAllRunningAgentsForTask(taskID string, role Role) []*Agent {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-	var result []*Agent
-	for _, a := range m.agents {
-		if a.TaskID != taskID || !isLive(a.GetState()) {
-			continue
-		}
-		if role != "" && RoleFromName(a.Name) != role {
-			continue
-		}
-		result = append(result, a)
-	}
-	return result
-}
-
-// KillAgentsForTask stops all running agents for the given task ID and waits
-// for their goroutines to exit (up to timeout). Safe to call from DeleteTask
-// before worktree cleanup.
-func (m *Manager) KillAgentsForTask(taskID string, timeout time.Duration) {
-	m.mu.RLock()
-	var targets []*Agent
-	for _, a := range m.agents {
-		if a.TaskID == taskID {
-			targets = append(targets, a)
-		}
-	}
-	m.mu.RUnlock()
-
-	for _, a := range targets {
-		m.logger.Info("agent.kill-for-task", "agent_id", a.ID, "task_id", taskID)
-		if a.cancel != nil {
-			a.cancel()
-		}
-		a.SetState(StateStopped)
-		a.stdinMu.Lock()
-		if a.stdinPipe != nil {
-			_ = a.stdinPipe.Close()
-			a.stdinPipe = nil
-		}
-		a.stdinMu.Unlock()
-		m.emit(events.AgentState(a.ID), a)
-	}
-
-	deadline := time.After(timeout)
-	for _, a := range targets {
-		if a.done == nil {
-			continue
-		}
-		select {
-		case <-a.done:
-		case <-deadline:
-			m.logger.Warn("agent.kill-timeout", "agent_id", a.ID, "task_id", taskID)
-			return
-		}
-	}
-}
-
-func (m *Manager) StopAgent(agentID string) error {
-	m.mu.Lock()
-	a, ok := m.agents[agentID]
-	m.mu.Unlock()
-	if !ok {
-		return fmt.Errorf("agent %s not found", agentID)
-	}
-
-	m.logger.Info("agent.stop", "id", agentID)
-
-	a.MarkStopped()
-	// Send SIGINT first so CC can restore terminal modes and persist the
-	// session ID for --resume. Escalate to SIGKILL only after the grace window.
-	if a.Mode == "headless" {
-		if cmd := a.GetCmd(); cmd != nil {
-			stopWithSIGINT(cmd, a.done, stopSIGINTGrace)
-		}
-	}
-	if a.cancel != nil {
-		a.cancel()
-	}
-	a.SetState(StateStopped)
-	// Close stdin to signal the claude process to exit.
-	a.stdinMu.Lock()
-	if a.stdinPipe != nil {
-		_ = a.stdinPipe.Close()
-		a.stdinPipe = nil
-	}
-	a.stdinMu.Unlock()
-	m.emit(events.AgentState(agentID), a)
-	return nil
-}
-
-func (m *Manager) GetAgent(agentID string) (*Agent, error) {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-	a, ok := m.agents[agentID]
-	if !ok {
-		return nil, fmt.Errorf("agent %s not found", agentID)
-	}
-	return a, nil
-}
-
-func (m *Manager) ListAgents() []*Agent {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-	agents := make([]*Agent, 0, len(m.agents))
-	for _, a := range m.agents {
-		agents = append(agents, a)
-	}
-	return agents
-}
-
-// HasRunningAgentForTask returns true if any agent is currently running for the given task.
-// For headless agents this checks whether the goroutine has truly exited (via the done
-// channel) rather than the State field, which may be set to Stopped by StopAgent before
-// the goroutine finishes — avoiding a race where the worktree is cleaned up while the
-// agent process is still using it.
-func (m *Manager) HasRunningAgentForTask(taskID string) bool {
-	m.mu.RLock()
-	defer m.mu.RUnlock()
-	for _, a := range m.agents {
-		if a.TaskID != taskID {
-			continue
-		}
-		if a.done != nil {
-			// headless: goroutine alive until done is closed
-			select {
-			case <-a.done:
-				// goroutine exited
-			default:
-				return true
-			}
-		} else if a.GetState() == StateRunning {
-			// interactive: no goroutine, rely on state
-			return true
-		}
-	}
-	return false
-}
-
-// defaultShutdownGrace bounds how long Shutdown waits for in-flight agents
-// to notice the cancel signal and exit cleanly. Without this, agents get
-// SIGKILL'd mid-stream and the final `result` NDJSON line is truncated,
-// leaving operators with a half-written run log and a "signal: killed"
-// exit error that carries no diagnostic value.
-const defaultShutdownGrace = 20 * time.Second
-
-// Shutdown cancels all running agents and blocks until they exit or the
-// grace window elapses — whichever comes first. Subprocess runners use
-// SIGTERM + cmd.WaitDelay (see shutdownSignal in runner helpers) so this
-// grace gives claude/codex time to flush the final `result` event.
-func (m *Manager) Shutdown() {
-	m.ShutdownWithGrace(defaultShutdownGrace)
-}
-
-// ShutdownWithGrace is Shutdown with an explicit grace window. Used by
-// tests (short grace) and callers that want to override the default.
-func (m *Manager) ShutdownWithGrace(grace time.Duration) {
-	m.mu.RLock()
-	count := len(m.agents)
-	dones := make([]chan struct{}, 0, count)
-	for _, a := range m.agents {
-		if a.cancel != nil {
-			a.cancel()
-		}
-		if a.done != nil {
-			dones = append(dones, a.done)
-		}
-	}
-	m.mu.RUnlock()
-
-	m.logger.Info("agent.shutdown", "count", count, "grace", grace, "wait", len(dones))
-	if len(dones) == 0 || grace <= 0 {
-		return
-	}
-
-	deadline := time.After(grace)
-	for i, done := range dones {
-		select {
-		case <-done:
-		case <-deadline:
-			m.logger.Warn("agent.shutdown.timeout",
-				"exited", i, "remaining", len(dones)-i, "grace", grace)
-			return
-		}
-	}
-	m.logger.Info("agent.shutdown.done", "exited", len(dones))
-}
-
-// safeArgRe matches only characters safe to embed in a shell command
-// without quoting: alphanumerics, dot, underscore, hyphen, forward-slash.
-var safeArgRe = regexp.MustCompile(`^[a-zA-Z0-9._/-]+$`)
